refactor: endpoint 이름 수정 및 기능 재구조화
- endpoint들을 restful하게 바꿈: 기능이 명시적이게
- folder에서 장소 정보들 보여주는 기능을 place가 아닌 folder controller에서 처리함.
- 장소 정보 추출을 user Repo가 아닌 folder_place Repo에서 수행함. : 반환할 장소 데이터들이 folder_place에 저장된 장소들이기 때문.
