+++ src/place/place.controller.ts
@@ -1,28 +1,8 @@
-import {
-  Body,
-  Controller,
-  Get,
-  Param,
-  Post,
-  Query,
-  Req,
-  UseGuards,
-} from '@nestjs/common';
+import { Body, Controller, Get, Param, Post } from '@nestjs/common';
 import { PlaceService } from './place.service';
-import {
-  ApiBearerAuth,
-  ApiBody,
-  ApiOperation,
-  ApiQuery,
-  ApiResponse,
-  ApiTags,
-} from '@nestjs/swagger';
+import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { SearchByTextReqDto } from './dtos/search-by-text-req.dto';
-import { AuthGuard } from '@nestjs/passport';
-import { MarkerResDto } from './dtos/marker-res.dto';
 import { PlacePreviewResDto } from './dtos/place-preview-res.dto';
-import { PlaceListResDto } from './dtos/place-list-res.dto';
-import { PlaceListReqDto } from './dtos/place-list-req.dto';
 
 @ApiTags('places')
 @Controller('places')
@@ -71,83 +51,15 @@ export class PlaceController {
   //   return await this.placeService.createPlaceTag(createPlaceTagReqDto);
   // }
 
-  @UseGuards(AuthGuard('access'))
-  @ApiBearerAuth('Access Token')
-  @ApiOperation({ summary: "Get User's place markers" })
-  @ApiResponse({ type: MarkerResDto })
-  @ApiQuery({ name: 'addressList', required: false, type: [String] })
-  @ApiQuery({ name: 'categoryList', required: false, type: [String] })
-  @Get('/markers/all')
-  async getAllMarkers(
-    @Query('addressList') addressList: string[] = [],
-    @Query('categoryList') categoryList: string[] = [],
-    @Req() req,
-  ) {
-    return await this.placeService.getMarkers(
-      req.user.id,
-      addressList,
-      categoryList,
-    );
-  }
-
-  @UseGuards(AuthGuard('access'))
-  @ApiBearerAuth('Access Token')
-  @ApiOperation({ summary: "Get User's place markers by folder" })
-  @ApiResponse({ type: MarkerResDto })
-  @ApiQuery({ name: 'addressList', required: false, type: [String] })
-  @ApiQuery({ name: 'categoryList', required: false, type: [String] })
-  @Get('/markers/folder/:id')
-  async getMarkersByFolder(
-    @Param('id') folderId: number,
-    @Query('addressList') addressList: string[] = [],
-    @Query('categoryList') categoryList: string[] = [],
-    @Req() req,
-  ) {
-    return await this.placeService.getMarkers(
-      req.user.id,
-      addressList,
-      categoryList,
-      folderId,
-    );
-  }
-
   @ApiOperation({ summary: "Get place's preview info from marker" })
   @ApiResponse({ type: PlacePreviewResDto })
-  @Get('/place-preview/:id')
+  @Get('/:id/preview')
   async getPlaceInfoFromMarker(
     @Param('id') id: number,
   ): Promise<PlacePreviewResDto> {
     return await this.placeService.getPlaceInfoFromMarker(id);
   }
 
-  @UseGuards(AuthGuard('access'))
-  @ApiBearerAuth('Access Token')
-  @ApiOperation({ summary: "Get User's place list" })
-  @ApiResponse({ type: PlaceListResDto })
-  @Get('/list/all')
-  async getAllPlaceList(
-    @Req() req,
-    @Query() placeListReqDto: PlaceListReqDto,
-  ): Promise<PlaceListResDto> {
-    return this.placeService.getPlaceList(req.user.id, placeListReqDto);
-  }
-
-  @UseGuards(AuthGuard('access'))
-  @ApiBearerAuth('Access Token')
-  @ApiOperation({ summary: "Get User's place list by folder" })
-  @ApiResponse({ type: PlaceListResDto })
-  @Get('/list/folder/:id')
-  async getPlaceListByFolder(
-    @Req() req,
-    @Param('id') folderId: number,
-    @Query() placeListReqDto: PlaceListReqDto,
-  ): Promise<PlaceListResDto> {
-    return this.placeService.getPlaceList(
-      req.user.id,
-      placeListReqDto,
-      folderId,
-    );
-  }
   // @ApiOperation({ summary: 'Create placeImage' })
   // @Post('place-images')
   // async createPlaceImage(
